move Category from components to views
